pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolov8-api"
version = "0.1.0"
description = "YOLOv8 Object Detection API (FastAPI + Docker)"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "gunicorn",
  "prometheus-client",
  "pydantic-settings",
  "httpx",
  "pillow",
  "requests",
  "ultralytics",
  "opencv-python-headless",
  "torch; platform_system=='Linux'",
  "python-multipart",
  "pytest",
  "pytest-asyncio",
]


[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
